Clean code - Extract CostCalculator as an interface. This wil enable us to provide alternative cost calculators later.
